# docs/openapi.yaml
openapi: '3.0.2'
info:
    title: Collective IAM
    version: '1.0.0'
servers:
    - url: https://ciam.centralmind.net
paths:
...
    /user:
        post:
            security:
                - bearerAuth:
                      ['ciam.user.create', 'ciam.permission.grant.{FLAGS}']
            summary: 'Create a user'
            description: |
                Create a new role from at least a name.<br>
                `ciam.permission.grant.FLAGS` is only checked if permissions are provided, for example if you try to
                create a role with the permissions [`chipware.ban`, `chipware.unban`] then you will need to have the
                permissions [`ciam.permission.grant.chipware.ban`, `ciam.permission.grant.chipware.unban`].
            tags:
                - user
            requestBody:
                description: 'The user to create'
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    description: 'The name of this user'
                                roles:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/ObjectId'
                                    description: 'The role ids this user has'
                                permissions:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/Flag'
                                    description: 'The permission flags this user has'
                                discord:
                                    type: object
                                    properties:
                                        id:
                                            $ref: '#/components/schemas/DiscordId'
                                            description: 'User discord id'
                            required:
                                - name
            responses:
                '200':
                    description: 'Returns the newly created user'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                '400':
                    description: 'Invalid body, missing role given'
                '401':
                    description: 'Invalid token or missing permissions'
                '500':
                    description: 'Failed saving user to database'

        patch:
            summary: 'Update a user'
            security:
                - bearerAuth:
                      [
                          'ciam.user.update.{USER_ID}',
                          'ciam.permission.grant.{FLAGS}',
                          'ciam.permission.revoke.{FLAGS}',
                      ]
            tags:
                - user
            requestBody:
                description: |
                    The user to update, omitted fields wont be updated.<br>
                    `ciam.permission.grant|revoke.FLAGS` is only checked if permissions are provided, for example if you try to
                    set the permissions to [`chipware.ban`, `chipware.unban`] on a user that already has the permissions
                    [`chipware.ban`, `chipware.kick`] then you will need to have the permissions
                    [`ciam.permission.grant.chipware.unban`, `ciam.permission.revoke.chipware.kick`].
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                _id:
                                    $ref: '#/components/schemas/ObjectId'
                                    description: 'The id of the user to update'
                                name:
                                    type: string
                                    description: 'New name of this user'
                                permissions:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/Flag'
                                    description: 'New permissions of this user'
                                roles:
                                    type: array
                                    items:
                                        $ref: '#/components/schemas/ObjectId'
                                    description: 'New roles of this user'
                            required:
                                - _id
            responses:
                '200':
                    description: 'Returns the updated user'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                '400':
                    description: 'Invalid body'
                '401':
                    description: 'Invalid token or missing permissions'
                '404':
                    description: 'User not found'
                '500':
                    description: 'Failed saving user to database'
...
    /login/identify:
        get:
            summary: 'Simple OAuth endpoint'
            description: |
                Redirect users to this endpoint to perform a simple version of OAuth.
                Once the flow is complete the user will be redirected to your website with a code. You can then use this code to get the user info.
            tags:
                - auth
            parameters:
                - in: query
                  name: redirect
                  description: |
                      The url to redirect the user to. The user will have the code as a query parameter.
                      For example: If the redirect url is 'centralmind.net' then the user will be redirected to 'centralmind.net?code=THEIR_CODE'
                  required: false
                  schema:
                      type: string
            responses:
                '302':
                    description: 'Redirect to discord OAuth2'
    /auth/identify/{code}:
        get:
            summary: 'Get user information'
            description: |
                Get user information from a code after they have completed the indentification flow.
            tags:
                - auth
            parameters:
                - in: path
                  name: code
                  required: true
                  schema:
                      $ref: '#/components/schemas/uuid'
                  description: 'The code returned when the user was redirected back to your website'
            responses:
                '200':
                    description: 'Returns the user data'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    code:
                                        $ref: '#/components/schemas/uuid'
                                        description: 'One time code for getting this data'
                                    redirect:
                                        type: string
                                        description: |
                                            Url the code for this data was redirected token.
                                            Make sure this matches the redirect url you provided.
                                    user:
                                        type: object
                                        description: 'Partial user object'
                                        properties:
                                            _id:
                                                $ref: '#/components/schemas/ObjectId'
                                                description: 'The users ID, not the same as a discord ID'
                                            name:
                                                type: string
                                                description: 'The name of this user'
                                            avatar:
                                                type: string
                                                description: 'Avatar url'
                                            discord:
                                                $ref: '#/components/schemas/UserDiscord'
                                        required:
                                            - _id
                                            - name
                                            - avatar
                                            - discord
                                required:
                                    - code
                                    - redirect
                                    - user

                '401':
                    description: 'Invalid token or missing permissions'
                '400':
                    description: 'Invalid code or user no longer exists'
    /login:
        get:
            summary: 'Redirect endpoint for OAuth'
            description: |
                Once the user has completed the Oauth2 flow they will be redirected to the `returnUrl`, they will have a cookie
                set for the domain `domain`, this cookie can be used to access the users information.
            tags:
                - auth
            parameters:
                - in: query
                  name: returnUrl
                  description: 'The url to return to after the user has completed oauth'
                  required: true
                  schema:
                      # TODO: Create a schema for urls?
                      type: string
                - in: query
                  name: domain
                  description: 'The domain to set the auth cookie for, this should be your websites domain'
                  required: true
                  schema:
                      # TODO: Create a schema for urls?
                      type: string
                - in: query
                  name: scope
                  description: 'The scope you are requesting for a user, currently only "identify"'
                  required: true
                  schema:
                      # TODO: Create a schema for scopes?
                      type: string
                - in: query
                  name: client
                  description: 'Client id of your application'
                  required: true
                  schema:
                      $ref: '#/components/schemas/ObjectId'
            responses:
                '302':
                    description: 'Redirect to discord OAuth2'
                '400':
                    description: 'Bad request'
    /auth/user:
        get:
            summary: 'Get user'
            description: |
                Get a user from the access token.
            tags:
                - auth
            parameters:
                - in: header
                  name: ciam_token
                  description: 'The access token'
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: 'Redirect to discord OAuth2'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        $ref: '#/components/schemas/ObjectId'
                                        description: 'User id'
                                    username:
                                        type: string
                                        description: 'Username'
                                    discord:
                                        $ref: '#/components/schemas/UserDiscord'
                '400':
                    description: 'Bad request'
                '401':
                    description: 'Invalid token or missing permissions'
    /auth/refresh:
        get:
            summary: 'Refresh access token'
            description: |
                Refresh an access token for a user.
            tags:
                - auth
            parameters:
                - in: query
                  name: refreshToken
                  description: 'The refresh token'
                  required: true
                  schema:
                      type: string
            responses:
                '302':
                    description: 'Redirect to discord OAuth2'
                '400':
                    description: 'Bad request'
                '500':
                    description: 'Error creating new token'
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        uuid:
            type: string
            pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
            # the regex above limits the length;
            # however, some tools might require explicit settings:
            minLength: 36
            maxLength: 36
        ObjectId:
            type: string
            maxLength: 24
            pattern: '^[0-9a-f]{24}$'
        DiscordId:
            type: string
            minLength: 16
            maxLength: 20
            pattern: '^[0-9]{16,20}$'
        Flag:
            type: string
            minLength: 1
            pattern: '^(([a-z]+|\?)(\.([a-z]+|\?))*(\.\*)?)|\*$'
...
        UserDiscord:
            type: object
            description: 'User discord data'
            properties:
                id:
                    $ref: '#/components/schemas/DiscordId'
                    description: 'Discord id'
                name:
                    type: string
                    description: 'Discord username'
                discriminator:
                    type: string
                    description: 'Discord discriminator'
            required:
                - id
        User:
            type: object
            description: 'A user can have both permissions and roles, and can be tied to a discord user, or be standalone'
            properties:
                _id:
                    $ref: '#/components/schemas/ObjectId'
                    description: 'The users ID, not the same as a discord ID'
                name:
                    type: string
                    description: 'The name of this user'
                avatar:
                    type: string
                    description: 'Avatar url'
                roles:
                    type: array
                    items:
                        $ref: '#/components/schemas/ObjectId'
                    description: 'List of role ids this user has'
                permissions:
                    type: array
                    items:
                        $ref: '#/components/schemas/Flag'
                    description: 'List of permission flags this user has'
                discord:
                    $ref: '#/components/schemas/UserDiscord'
            required:
                - _id
                - name
                - avatar
                - roles
                - permissions
...